ux/client: the handlers for clearing the {measure} layer pixel path and its selection now live with the  other mutators
